added component for slider selection and moved beverage calculation imports to scale component
